+++ esbuild.js
@@ -1,9 +1,16 @@
 const esbuild = require("esbuild");
 
+/* modules for web extensions */
+const glob = require('glob');
+const path = require('path');
+const polyfill = require('@esbuild-plugins/node-globals-polyfill');
+
 const production = process.argv.includes('--production');
 const watch = process.argv.includes('--watch');
 
 /**
+ * This plugin hooks into the build process to print errors in a format that the problem matcher in
+ * Visual Studio Code can understand.
  * @type {import('esbuild').Plugin}
  */
 const esbuildProblemMatcherPlugin = {
@@ -23,8 +30,37 @@ const esbuildProblemMatcherPlugin = {
 	},
 };
 
+/**
+ * For web extension, all tests, including the test runner, need to be bundled into
+ * a single module that has a exported `run` function .
+ * This plugin bundles implements a virtual file extensionTests.ts that bundles all these together.
+ * @type {import('esbuild').Plugin}
+ */
+const testBundlePlugin = {
+	name: 'testBundlePlugin',
+	setup(build) {
+		build.onResolve({ filter: /[\/\\]extensionTests\.ts$/ }, args => {
+			if (args.kind === 'entry-point') {
+				return { path: path.resolve(args.path) };
+			}
+		});
+		build.onLoad({ filter: /[\/\\]extensionTests\.ts$/ }, async args => {
+			const testsRoot = path.join(__dirname, 'src/test');
+			// const testsRoot = path.join(__dirname, 'src/web/test/suite');
+			const files = await glob.glob('*.test.{ts,tsx}', { cwd: testsRoot, posix: true });
+			return {
+				contents:
+					`export { run } from './mochaTestRunner.ts';` +
+					files.map(f => `import('./${f}');`).join(''),
+				watchDirs: files.map(f => path.dirname(path.resolve(testsRoot, f))),
+				watchFiles: files.map(f => path.resolve(testsRoot, f))
+			};
+		});
+	}
+};
+
 async function main() {
-	const ctx = await esbuild.context({
+	const nodeCtx = await esbuild.context({
 		entryPoints: [
 			'src/extension.ts'
 		],
@@ -34,7 +70,7 @@ async function main() {
 		sourcemap: !production,
 		sourcesContent: false,
 		platform: 'node',
-		outfile: 'dist/extension.js',
+		outfile: 'dist/node/extension.js',
 		external: ['vscode'],
 		logLevel: 'silent',
 		plugins: [
@@ -42,11 +78,45 @@ async function main() {
 			esbuildProblemMatcherPlugin,
 		],
 	});
+
+	const webCtx = await esbuild.context({
+		entryPoints: [
+			'src/extension.ts',
+			'src/test/extensionTests.ts'
+		],
+		bundle: true,
+		format: 'cjs',
+		minify: production,
+		sourcemap: !production,
+		sourcesContent: false,
+		platform: 'browser',
+		outdir: 'dist/web',
+		external: ['vscode'],
+		logLevel: 'silent',
+		// Node.js global to browser globalThis
+		define: {
+			global: 'globalThis',
+		},
+
+		plugins: [
+			polyfill.NodeGlobalsPolyfillPlugin({
+				process: true,
+				buffer: true,
+			}),
+			testBundlePlugin,
+			esbuildProblemMatcherPlugin, /* add to the end of plugins array */
+		],
+	});
+
 	if (watch) {
-		await ctx.watch();
+		await nodeCtx.watch();
+		await webCtx.watch();
+		// await Promise.all([nodeCtx.watch(), webCtx.watch()]);
 	} else {
-		await ctx.rebuild();
-		await ctx.dispose();
+		await nodeCtx.rebuild();
+		await nodeCtx.dispose();
+		await webCtx.rebuild();
+		await webCtx.dispose();
 	}
 }